refactor: split prompt routes into endpoint modules
