+++ routes/prompts/blocks/__init__.py
@@ -1,0 +1,14 @@
+from .helpers import router, supabase
+
+from . import get_blocks
+from . import get_blocks_by_type
+from . import create_block
+from . import update_block
+from . import delete_block
+from . import get_block_types
+from . import seed_sample_blocks
+
+__all__ = [
+    "router",
+    "supabase",
+]
